finished server

+++ post/api/views.py
@@ -1,5 +1,6 @@
 import json
 
+from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.permissions import IsAuthenticated, AllowAny
 
 from post.models import *
@@ -30,15 +31,25 @@ def post_all(request):
 @permission_classes((IsAuthenticated,))
 def post_create(request):
     current_user = request.user
-    req_body = request.body
-    js = json.loads(req_body)
-    post = Post(title=js['title'], description=js['description'], uploader=current_user)
-    if 'image' in request.data.keys():
-        post.image = request.data['image']
-    post.time = datetime.now()
-    post.save()
-    serializer = PostSerializer(post)
-    return Response(serializer.data, status=status.HTTP_201_CREATED)
+    try:
+        if 'file' in request.data.keys():
+            js = request.data
+            print(js)
+            post = Post(title=js['title'], description=js['description'], uploader=current_user, image=js['file'])
+            post.time = datetime.now()
+            post.save()
+            print(post.image.path)
+            serializer = PostSerializer(post)
+            return Response({'status': 'Successfully added post'}, status=status.HTTP_201_CREATED)
+        else:
+            js = request.data
+            post = Post(title=js['title'], description=js['description'], uploader=current_user)
+            post.time = datetime.now()
+            post.save()
+            # serializer = PostSerializer(post)
+            return Response({'status': 'Successfully added post'}, status=status.HTTP_201_CREATED)
+    except Exception as e:
+        return Response({'status': f'Failed because of: \n{e}'}, status=status.HTTP_201_CREATED)
 
 
 @api_view(['DELETE', ])
@@ -50,31 +61,43 @@ def post_delete(request, id):
     if count == 0:
         return Response({'status': f'Post with ID: {id} not found'}, status=status.HTTP_400_BAD_REQUEST)
     post = Post.objects.get(pk=id)
-    post.delete()
+    if request.user.pk == post.uploader.pk:
+        post.delete()
+    else:
+        return Response({'status': f'You are not the post creator'}, status=status.HTTP_403_FORBIDDEN)
+    print('----------------')
     return Response({'status': f'Successfully deleted the post'}, status=status.HTTP_200_OK)
 
 
 @api_view(['PUT', ])
 @permission_classes((IsAuthenticated,))
 def post_update(request, id):
+    data = request.data
     if id is None:
         return Response({'status': 'No ID sent'}, status=status.HTTP_400_BAD_REQUEST)
     count = Post.objects.filter(pk=id).count()
     if count == 0:
         return Response({'status': f'Post with ID: {id} not found'}, status=status.HTTP_400_BAD_REQUEST)
     post = Post.objects.get(pk=id)
-    req_body = request.body
-    js = json.loads(req_body)
-    if js['title'] is not None:
-        post.title = js['title']
-    if js['description'] is not None:
-        post.description = js['description']
-    if 'image' in request.data.keys():
-        post.image = request.data['image']
+    if 'file' in data.keys():
+        new_file = request.FILES.get('file')
+        # print(js)
+        if data['title'] is not None:
+            post.title = data['title']
+        if data['description'] is not None:
+            post.description = data['description']
+        print('----------------------------------------------------------')
+        post.image = new_file
+        post.time = datetime.now()
+        post.save()
+        return Response({'status': 'updated successfully'}, status=status.HTTP_201_CREATED)
+    if data['title'] is not None:
+        post.title = data['title']
+    if data['description'] is not None:
+        post.description = data['description']
     post.time = datetime.now()
     post.save()
-    serializer = PostSerializer(post)
-    return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response({'status': 'updated successfully'}, status=status.HTTP_200_OK)
 
 
 @api_view(['GET', ])
@@ -113,10 +136,12 @@ def register(request):
             return Response({'status': 'No password sent of Bad password sent'}, status=status.HTTP_400_BAD_REQUEST)
 
         new_user = User(first_name=user['first_name'], last_name=user['last_name'], email=user['email'],
-                        password=user['password'], username=user['username'], )
+                        username=user['username'])
+        new_user.set_password(user['password'])
         new_user.save()
         token, _ = Token.objects.get_or_create(user=new_user)
         user['token'] = token.key
+        user['id'] = new_user.pk
 
     except Exception as e:
         return Response({'status': str(e)}, status=status.HTTP_400_BAD_REQUEST)
@@ -143,7 +168,7 @@ def login(request):
     data['user_name'] = user.username
     data['email'] = user.email
     data['token'] = token.key
-    data['password'] = password
+    data['id'] = user.pk
     js = json.dumps(data, indent=None)
     return Response(json.loads(js),
                     status=status.HTTP_200_OK)
